requirements.txt
fastapi
uvicorn[standard]
pydantic
pyswisseph
pytz
timezonefinder
matplotlib
numpy
fastapi-cache2[redis]
gunicorn